rtmpserver: try to retain fmt from publisher to subscribers
FMT or "Chunk Message Header Format" is used to describe the header,
where 0 is the most basic (12 bytes), and 3 is the smallest (1 byte).

For audio, which shares a lot of data for each packet, there is lots of
bandwidth to be saved by using mostly type 3, since things like
timestamp has a constant delta (20ms), and only type 0 deals with absolute
timestamp, all the others (1, 2, 3) deals with deltas.

The problem with fmt is that type 3 is also used when a message is broken
into several parts (chunks), so that with a chunk-size of 128, and a
message-size of 200, you will get one chunk of 128, followed by a
type 3 header and then another 72 bytes. Now, since this is tcp, all
this data might arrive at the same time at the receiver, so then it
is very important to know which chunk-size the sender is using so
that we can correctly parse the data, since a differing chunk-size at
sender and receiver would potentially disguise the fmt-header byte in the
bitstream, causing AMF parsing-failures.

Now, we need to ignore the senders type 3 fmt to get the whole message,
as well as adding our own type 3 fmt for sending when breaking up
the message in appropriate chunks, but apart from that, other fmt's
including type 3 that can come as a result of audio-timestamp-deltas
being the same (speex, 20ms), we now try to forward to the subscribers.