rtmpserver: send returning 0 is allowed
...and passing range=0 to g_byte_array_remove_range is fatal.

Ref: http://stackoverflow.com/questions/3081952/with-c-tcp-sockets-can-send-return-zero

No explicit test for this as it is apparently a very rare condition for
a TCP socket, and required a whole bunch of stuff that would be very
hard to (but maybe not impossible?) to do. For now the pexcellent-RTMP
can be seen as the ones keeping this from happening once in a blue moon.

Fixes #3436