RTMPServer: clear out OpenSSL error queue after loading (EC)DH parameters.
Corollary to the equivalent change in Python. As we don't
consider the error queue at all in RTMPServer (except for
printing it out when we've detected an error by some other
means) this is harmless. The net effect is to remove the
expected errors from logging, which should make some things
simpler to debug.