rtmp: support Flash-flavor of extended timestamp
Our flavor of extended timestamp was extracted from the published RTMP
spec, but as this link points out (thanks David)
https://forums.adobe.com/thread/542749?tstart=0

...Flash does something completely different, namely embedding the
extended timestamp after each chunk in a larger-then-chunk message,
(something that is completely stupid to do, since you use much more
bytes for no reason whatsoever!). However, if you are not adhering to
this, a receiver familiar with flash will (most likely) treat the
first 4 bytes of your message as an extended timestamp,
skewing all data parsing, and creating certain havoc. Now a smart receiver
could check if the 4 bytes after the fmt-3 byte is indeed matching the
extended timestamp for this message, but I doubt any receivers are that
smart, given that we have failed both against Wowza, Flash and YouTube.
(And I have not implemented that in our server yet either)

Fixes #4818

Also, since a lot of test-infrastructure had to be restructured to allow
testing this properly, a slight bug was found relating to blocksize for
the rtmpsrc. I had thought that setting that to our rtmp-chunk size would
be the right fix, but if you have 128 (default) chunk size for rtmp, and
a packet of length 129, blocksize 128 will first send the first 128 bytes,
but will then wait until it has another 127 bytes before sending the next
one to flvdemux, so the only right number (for testing) is 1! Since
this will cause the tests to be more undeterministic then needed, I am
going to punt on this Fixes #4378 :)