server: remove locking
By extending the dialout-client principle of always keeping work to the
rtmp-thread, with direct-access we now use an atomic queue when adding
a client, and combined with some simple ref-counting of the client,
ensures that both the "outside" API surface of the rtmpserver as well
as the "inside" with the rtmp-thread can guarantee the client being
alive and well.

The notification-signals has also gotten some love, where we will now
notify of the direct-connection, *unless* the "ignore-localhost" flag
is set (needs a new name!), nicely replicating the old behavior when
rtmpsrc and sink was running on localhost to interact with the server.